+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: nouakhro <nouakhro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/02 14:47:38 by nouakhro          #+#    #+#             */
-/*   Updated: 2023/05/04 15:28:30 by nouakhro         ###   ########.fr       */
+/*   Updated: 2023/05/07 16:30:11 by nouakhro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,48 +21,48 @@ typedef struct s_list
 	struct s_list	*next;
 }	t_list;
 
-size_t	ft_strlcpy(char *dst, const char *src, size_t size);
-size_t	ft_strlcat(char *dst, const char *src, size_t size);
-size_t	ft_strlen(const char	*str);
-int		ft_isalnum(int argument);
-int		ft_isalpha(int argument);
-t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-int		ft_isascii(int a);
-int		ft_isdigit(int a);
-int		ft_toupper(int ch);
-int		ft_tolower(int character);
-int		ft_strncmp(const char *frst, const char *scnd, size_t length);
-int		ft_memcmp(const void *str1, const void *str2, size_t n);
-int		ft_atoi(const char *str);
-int		ft_isprint(int arg);
-void	ft_striteri(char *s, void (*f)(unsigned int, char*));
-void	ft_putchar_fd(char c, int fd);
-void	ft_putnbr_fd(int n, int fd);
-void	ft_putstr_fd(char *s, int fd);
-void	ft_putendl_fd(char *s, int fd);
-void	*ft_calloc(size_t count, size_t size);
-void	*ft_memchr(const void *str, int c, size_t n);
-void	*ft_memcpy(void *dst, const void *src, size_t n);
-void	*ft_memset(void *str, int c, size_t n);
-void	*ft_memmove(void *str1, const void *str2, size_t n);
-void	ft_bzero(void *s, size_t n);
-void	ft_lstdelone(t_list *lst, void (*del)(void*));
-void	ft_lstclear(t_list **lst, void (*del)(void*));
-int		ft_lstsize(t_list *lst);
-void	ft_lstiter(t_list *lst, void (*f)(void *));
-void	ft_lstadd_back(t_list **lst, t_list *new);
-t_list	*ft_lstlast(t_list *lst);
-t_list	*ft_lstnew(void *content);
-char	*ft_strchr(const char *str, int chr);
-char	*ft_strrchr(const char *str, int chr);
-char	*ft_strnstr(const char *s1, const char *s2, size_t len);
-char	*ft_strdup(const char *s1);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
-char	*ft_strjoin(char const *s1, char const *s2);
-char	*ft_itoa(int n);
-char	*ft_strtrim(char const *s1, char const *set);
-char	**ft_split(char const *s, char c);
-char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-void	ft_lstadd_front(t_list **lst, t_list *new);
+size_t		ft_strlcpy(char *dst, const char *src, size_t size);
+size_t		ft_strlcat(char *dst, const char *src, size_t size);
+size_t		ft_strlen(const char	*str);
+int			ft_isalnum(int argument);
+int			ft_isalpha(int argument);
+t_list		*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
+int			ft_isascii(int a);
+int			ft_isdigit(int a);
+int			ft_toupper(int ch);
+int			ft_tolower(int character);
+int			ft_strncmp(const char *frst, const char *scnd, size_t length);
+int			ft_memcmp(const void *str1, const void *str2, size_t n);
+long int	ft_atoi(const char *str);
+int			ft_isprint(int arg);
+void		ft_striteri(char *s, void (*f)(unsigned int, char*));
+void		ft_putchar_fd(char c, int fd);
+void		ft_putnbr_fd(int n, int fd);
+void		ft_putstr_fd(char *s, int fd);
+void		ft_putendl_fd(char *s, int fd);
+void		*ft_calloc(size_t count, size_t size);
+void		*ft_memchr(const void *str, int c, size_t n);
+void		*ft_memcpy(void *dst, const void *src, size_t n);
+void		*ft_memset(void *str, int c, size_t n);
+void		*ft_memmove(void *str1, const void *str2, size_t n);
+void		ft_bzero(void *s, size_t n);
+void		ft_lstdelone(t_list *lst, void (*del)(void*));
+void		ft_lstclear(t_list **lst, void (*del)(void*));
+int			ft_lstsize(t_list *lst);
+void		ft_lstiter(t_list *lst, void (*f)(void *));
+void		ft_lstadd_back(t_list **lst, t_list *new);
+t_list		*ft_lstlast(t_list *lst);
+t_list		*ft_lstnew(void *content);
+char		*ft_strchr(const char *str, int chr);
+char		*ft_strrchr(const char *str, int chr);
+char		*ft_strnstr(const char *s1, const char *s2, size_t len);
+char		*ft_strdup(const char *s1);
+char		*ft_substr(char const *s, unsigned int start, size_t len);
+char		*ft_strjoin(char const *s1, char const *s2);
+char		*ft_itoa(int n);
+char		*ft_strtrim(char const *s1, char const *set);
+char		**ft_split(char const *s, char c);
+char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
+void		ft_lstadd_front(t_list **lst, t_list *new);
 
 #endif
